BLD: simplify setup.py and reorder build folders
