+++ .readthedocs.yml
@@ -6,8 +6,8 @@ build:
 python:
   version: 3.6
   install:
-    - requirements: requirements.txt
-    - requirements: docs/requirements.txt
+    - requirements: requirements/requirements.txt
+    - requirements: requirements/requirements_docs.txt
     - method: setuptools
       path: .
   system_packages: true
