+++ pyproject.toml
@@ -28,21 +28,21 @@ exclude = '''
 # manylinyx1 vs manylinux2010 issues (for Equinor Komodo installs)
 [build-system]
 requires = [
-         "pip>=19.1.1",
-         "setuptools>=30.3.0",
-         "wheel",
-         "scikit-build",
-         'cmake==3.13.3; python_version < "3.7" and platform_system == "Linux"',
-         'cmake==3.15.3; python_version >= "3.7" and platform_system == "Linux"',
-         'cmake==3.18.0; platform_system != "Linux"',
-         "ninja",
-         "setuptools_scm>=3.2.0",
-         'numpy==1.10.4; python_version == "3.4"',
-         'numpy==1.13.3; python_version >= "3.5" and python_version < "3.7" and platform_system == "Linux"',
-         'numpy==1.16.0; python_version >= "3.7" or platform_system != "Linux"',
-         'Sphinx',
-         'sphinx-rtd-theme',
-         'sphinxcontrib-apidoc',
-         'sphinx-autodoc-typehints',
-         'recommonmark',
-         ]
+  "pip>=19.1.1",
+  "setuptools>=30.3.0",
+  "wheel",
+  "scikit-build",
+  'cmake==3.13.3; python_version < "3.7" and platform_system == "Linux"',
+  'cmake==3.15.3; python_version >= "3.7" and platform_system == "Linux"',
+  'cmake==3.18.0; platform_system != "Linux"',
+  "ninja",
+  "setuptools_scm>=3.2.0",
+  'numpy==1.10.4; python_version == "3.4"',
+  'numpy==1.13.3; python_version >= "3.5" and python_version < "3.7" and platform_system == "Linux"',
+  'numpy==1.16.0; python_version >= "3.7" or platform_system != "Linux"',
+  'Sphinx',
+  'sphinx-rtd-theme',
+  'sphinxcontrib-apidoc',
+  'sphinx-autodoc-typehints',
+  'recommonmark',
+]
