+++ requirements/requirements_dev.txt
@@ -1,0 +1,34 @@
+setuptools>=43
+setuptools_scm>=3.2.0
+scikit-build
+ninja>=1.9.0.post1
+cmake>3.13.3
+pip>=20
+wheel>=0.29.0
+flake8
+pydocstyle
+coverage>=4.1
+pytest-runner>=2.11.1
+pre-commit
+coverage>=4.1
+Sphinx
+sphinx-rtd-theme
+sphinxcontrib-apidoc
+sphinx-autodoc-typehints
+recommonmark
+bandit
+numpy>1.13.1
+pandas>=0.21.0
+segyio>=1.8.6
+matplotlib>=1.5
+scipy>=0.17
+shapely>=1.6.2; python_version < "3.8"
+shapely==1.7a2; python_version == "3.8"
+black
+autopep8
+pylint
+pytest>=6
+pytest-cov
+h5py>=3
+hdf5plugin>=2.3
+tables>=3.5
