+++ requirements/requirements_docs.txt
@@ -1,10 +1,12 @@
 pip>=19.1.1
 setuptools>=30.3.0
-cmake==3.13.3
+cmake>=3.13.3
 scikit-build
 ninja
 setuptools_scm
-sphinxcontrib-apidoc
+pydocstyle
+Sphinx
 sphinx-rtd-theme
+sphinxcontrib-apidoc
 sphinx-autodoc-typehints
 recommonmark
