+++ requirements/requirements_setup_extras.txt
@@ -1,0 +1,9 @@
+flake8
+pydocstyle
+coverage>=4.1
+pre-commit
+coverage>=4.1
+bandit
+autopep8
+pylint
+pytest-cov
