--- requirements_dev.txt
@@ -1,40 +1,0 @@
-setuptools>=30.3.0
-setuptools_scm>=3.2.0
-scikit-build
-ninja>=1.9.0.post1
-cmake==3.13.3; python_version < "3.7" and platform_system == "Linux"
-cmake==3.15.3; python_version >= "3.7" and platform_system == "Linux"
-cmake==3.15.3; platform_system != "Linux"
-pip>=19.1.1
-wheel>=0.29.0
-flake8
-pydocstyle
-coverage>=4.1
-pytest-runner>=2.11.1
-pre-commit
-coverage>=4.1
-Sphinx
-sphinx-rtd-theme
-sphinxcontrib-apidoc
-sphinx-autodoc-typehints
-recommonmark
-bandit
-numpy==1.10.4; python_version == "3.4"
-numpy==1.13.3; python_version >= "3.5" and python_version < "3.7"
-numpy>=1.16.0; python_version >= "3.7"
-pandas>=0.21.0; python_version != "3.4"
-pandas==0.20.3; python_version == "3.4"
-segyio==1.8.6; python_version == "3.4"
-segyio>=1.8.6; python_version != "3.4"
-matplotlib>=1.5
-scipy>=0.17
-shapely>=1.6.2; python_version < "3.8"
-shapely==1.7a2; python_version == "3.8"
-black
-autopep8
-pylint
-pytest>=2.9.2
-pytest-cov
-h5py>=3
-hdf5plugin>=2.3
-tables>=3.5.1
