--- scripts/.appveyor.yml
@@ -1,70 +1,0 @@
-version: 1.0.{build}
-
-configuration:
-  - Release
-
-os: Visual Studio 2017
-image: Visual Studio 2017
-
-platform:
-  - x64
-
-environment:
-  matrix:
-    - PYTHON: C:\Python36
-      PYTHON_VERSION: 36
-      TAG_SCENARIO: True
-    - PYTHON: C:\Python37
-      PYTHON_VERSION: 37
-      TAG_SCENARIO: True
-    - PYTHON: C:\Python38
-      PYTHON_VERSION: 38
-      TAG_SCENARIO: True
-
-matrix:
-  fast_finish: true
-
-for:
--
-  matrix:
-    only:
-      - TAG_SCENARIO: True
-  skip_non_tags: True
-
-install:
-  - ps: choco install -y --no-progress --allow-empty-checksums -r swig --version=4.0.1
-  - IF "%platform%" == "x64" SET PYTHON=%PYTHON%-x64
-  - SET PATH=%PYTHON%;%PYTHON%\Scripts;%PATH%
-  - SET PIP_DISABLE_PIP_VERSION_CHECK=1
-  - python -m pip install pip>=19.3.1 --upgrade --user
-
-  - echo %PYTHON_VERSION%
-  - python -m pip install -r scripts/requirements_appveyor.txt
-  - python -m pip install twine
-
-before_build:
-  - set generator="Visual Studio 15 2017"
-  - IF "%platform%" == "x64" set generator="Visual Studio 15 2017 Win64"
-  - git fetch --tags
-  - git clone --depth 1 https://github.com/equinor/xtgeo-testdata ../xtgeo-testdata
-
-
-build_script:
-  - python setup.py bdist_wheel -G %generator%
-  - ps: ls dist/*.whl
-  - ps: python -m pip install $(ls dist/*.whl | % {$_.FullName}) --force --no-deps
-  - python -m pip freeze --all
-  - python -m pip list
-
-test_script:
-  - pytest --disable-warnings
-
-# deploy_script:
-#   - echo "Upload to PYPI if tag..."
-#   - ps: |
-#         Write-Host "TAG: $env:APPVEYOR_REPO_TAG"
-#         If ($env:APPVEYOR_REPO_TAG -eq "false") { return }
-
-#         Write-Host "Uploading wheels to pypi"
-#         Invoke-Expression "twine upload -u jcrivenaes --password $env:PYPIPASSWD --skip-existing dist/*"
-#         If ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode) }
