--- scripts/.travis.yml
@@ -1,140 +1,0 @@
-language: python
-
-python:
-  - 3.6
-
-os:
-  - linux
-
-git:
-  depth: false
-
-addons:
-  apt:
-    packages:
-      - swig
-
-services: docker
-
-sudo: required
-
-env:
-  global:
-    - TRAVISRUN=true
-    - REPO_DIR=.
-    - UNICODE_WIDTH=32
-    - MB_ML_VER=1
-    - PLAT=x86_64
-    - MB_PYTHON_VERSION=$TRAVIS_PYTHON_VERSION
-    - VERSION=$(echo $TRAVIS_TAG)
-    - BUILD_DEPENDS=""
-    - CONFIG_PATH=scripts/config.sh
-    - TEST_DEPENDS="pytest"
-
-matrix:
-  fast_finish: true
-  exclude:
-    - python: 3.6
-  include:
-    - os: linux
-      env:
-        - MB_PYTHON_VERSION=2.7
-        - MB_ML_VER=1
-      if: (tag IS true)
-    - os: linux
-      env: MB_PYTHON_VERSION=3.5
-      if: (tag IS true)
-    - os: linux
-      env: MB_PYTHON_VERSION=3.6
-      if: (tag IS true)
-    - os: linux
-      env:
-        - MB_PYTHON_VERSION=3.7
-        - MB_ML_VER=2010
-      if: (tag IS true)
-    - os: linux
-      dist: xenial
-      sudo: true
-      env:
-        - MB_PYTHON_VERSION=3.8
-        - MB_ML_VER=2010
-        - TEST_DEPENDS="pytest segyio==1.9.1"
-      if: (tag IS true) # (type = pull_request) OR (tag IS true)
-
-    # - os: osx
-    #   language: generic
-    #   env: MB_PYTHON_VERSION=3.6
-    #   python: 3.6
-    #   if: tag IS present
-    # - os: osx
-    #   language: generic
-    #   env: MB_PYTHON_VERSION=3.7
-    #   python: 3.7
-    #   if: (type = pull_request) OR (tag IS true)
-    # - os: osx
-    #   language: generic
-    #   env: MB_PYTHON_VERSION=3.8
-    #   python: 3.8
-    #   if: tag IS true
-
-before_install:
-  - echo "PWD is $PWD"
-  - unset -f pushd
-  - unset -f popd
-  - git clone https://github.com/matthew-brett/multibuild
-  - source multibuild/common_utils.sh
-  - source multibuild/travis_steps.sh
-  - before_install
-
-install:
-  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
-        brew update && brew install swig;
-    fi
-  - pip install pip==19.3.1
-  - pip install bandit
-  - build_wheel $REPO_DIR $PLAT
-
-before_script:
-  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
-        brew link --overwrite python;
-    fi
-  - bandit -c bandit.yml -r src/xtgeo
-
-script:
-  - install_run $PLAT
-  - mv wheelhouse dist
-  - ls -la .
-  # for a given version, the test is ran outside the multibuild docker in order to run coverage
-  # see also scripts/config.sh
-  - if [[ "$MB_PYTHON_VERSION" == "3.6" && "$TRAVIS_OS_NAME" == "linux"  ]]; then
-        export TRAVISRUN=true;
-        git clone --depth 1 https://github.com/equinor/xtgeo-testdata ../xtgeo-testdata;
-        pip install -r requirements.txt;
-        pip install dist/*.whl --force;
-        pip install pytest pytest-cov;
-        pytest tests --disable-warnings --cov=xtgeo --cov-report=xml:xtgeocoverage.xml;
-    fi
-
-
-after_success:
-  - ls -la ${TRAVIS_BUILD_DIR}/dist
-  - ls -la .
-  # upload coverage report to codecov server
-  - if [[ -f "xtgeocoverage.xml" ]]; then
-        bash <(curl -s https://codecov.io/bash) -Z -c -f xtgeocoverage.xml;
-    fi
-
-# before_deploy:
-#   - pip install numpy
-
-# deploy:
-#   - provider: pypi
-#     skip_cleanup: true
-#     skip_upload_docs: true
-#     distributions: build
-#     user: statoil-travis
-#     password:
-#       secure: "fN4ZaDGqv6lWwRWWHgBawE5L6h+9Epkn6HfgBCMKpMA5WG8F0H2vxDrP8+HnsNf7eVhr4Z6lG2k0pzpuJb9BUtpknORYPocREBp7sDOmxY/a/Y+3Q+1g1dXY6JcLJW8uQTANQSvVTd8b1QMP1FxoIsdxxk+65j/Pn5vjFNgEJu8d7JyJR2GWQ4YwLmODITWqmM6QBbs6vMPtDh6VwvJuArHmj/v0phoIQ0rBXshS5hdA78+nsIYE6RSBYuonBSCA6wcQq8bIvl3xugVvvEbJ3iqFObA/i5Xar8RRgea7xvVhaB27oRIkzlyg2kWAVUVkfWCJLvinn7VcBgJrfEJlfiaz8yHaQWjouGMdYAjEKWmZKpjJEMjJ3YgPymLKTIOmX5wVQE1NYcQt3TANg/NECB/t99K9Faidith6G58GVeAIRt+T564Ocr5ZXUxfWF35akWsbH8gBcPJLl9LHOOtYy1Wp7xs4t2aRzz9X0uNuL0C3WRYd7wLL8/fhXwPL/4dJTeHsyNp52vLLJ4ulqZRiDVEjq+dMniTKRrXXxuvhNTkR2b4BWIAdUj7bxY7EE2vEEFVfx2X5CzyAF10pTrbPJ85AJauWS9i7Xv8yyi9L+VTGw8jnVvg/hFmN99o7nPGG+BovdTK0zrhzRCmWNrv3BqJiDSsxtxPws4GUedhWBc="
-#     on:
-#       tags: True
-#       repo: equinor/xtgeo
