--- scripts/requirements_appveyor.txt
@@ -1,13 +1,0 @@
-numpy>=1.10.4
-matplotlib>=1.5
-scipy>=0.17
-segyio>=1.8.6
-pandas==0.25.3
-shapely>=1.6.0
-six>=1.11
-cmake>=3.13.3
-scikit-build>=0.10.0
-ninja>=1.8.2.post2
-pytest>=2.9.2
-pytest-runner>=2.11.1
-setuptools_scm>=3.2.0
