+++ setup.cfg
@@ -1,4 +1,3 @@
-
 [flake8]
 exclude = docs/conf.py, src/xtgeo/cxtgeo/*.py, .eggs, _skbuild
 max-line-length = 88
@@ -10,11 +9,3 @@ match = (?!(test_|_)).*\.py
 
 [aliases]
 test = pytest
-
-[options]
-setup_requires =
-  setuptools_scm
-  pytest_runner
-  cmake
-  ninja
-  scikit-build
